+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Added the ability to use the latest USS encoding when saving an exiting file.
 - Removed the Cancel button and replaced it with X button on all dialogue pop-ups
 - Added toggle bar to show and hide File Explorer.
-- Added a prompt to confirm actions such as closing modified file and closing all opened files 
+
 
 
 ## `2.7.0`